dead projectile issues fixed (collision detection is still happening on them)
tried making shot a variable in player instead. Now it is still having
some issues

also turns out the shot variable on player was never disabled

currently when the shot is still in oldstates it glitches, otherwise it
is fine. why? (breaks movement too when this is the case kinda too)
need to fix the whole current frame thing (use subtractions instead),
because other players have frame numbers that are different (sort of
completed that in handleshot, but still doesnt seem to have fixed it,
find other issue)